@davemecha rightfully pointed out that it would make sense to stack API key and TLS cert requirements
So we now kill the old decision of separating auth and restrictions (which I originally did to enforce that
at least one access control mechanism would be in place for every consumer) and only have restrictions.
The TLS cert restriction should allow to just say "client needs to provide *any* cert that the server recognizes",
which maybe needs another checkbox on the form. We'll see.